fix: break circular dependency between CorrectionProposalDialog and Panels
Extract shared types, constants, and utility functions into
correction-proposal-shared.ts. The circular import (Dialog → Panels →
Dialog) caused undefined references after minification, crashing the
import screen with "i is not a function".

[tb-343]
